requirements.txt
anyio
asgiref
astroid
attrs
bcrypt
certifi
cffi
charset-normalizer
click
contextlib2
contextvars
coverage
dataclasses
ecdsa
fastapi
flake8
greenlet
h11
idna
immutables
importlib-metadata
iniconfig
isort
lazy-object-proxy
mccabe
packaging
passlib
platformdirs
pluggy
py
pyasn1
pycodestyle
pycparser
pydantic
pyflakes
pylint
pyparsing
pytest
pytest-cov
pytest-flake8
pytest-isort
python-jose
python-multipart
requests
rsa
six
sniffio
SQLAlchemy
SQLAlchemy-Utils
sqlite-checkforeignkeys
starlette
toml
tomli
typed-ast
typing_extensions
urllib3
uvicorn
wrapt
zipp
mysqlclient
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
PyYAML==6.0
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3